% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/com_scaling.r
\name{com_scaling}
\alias{com_scaling}
\title{Performs scaling of functional or phylogenetic structure of simulated community.}
\usage{
com_scaling(
  com,
  grain_size = 8,
  use_tr = T,
  abundance_weighted = T,
  scaling_iter = 100,
  be_verbose = T
)
}
\arguments{
\item{com}{A list, returned by evolve_com.}

\item{grain_size}{A number, size of community data matrix.}

\item{use_tr}{If True, uses trait distances, else uses phylogenetic distances.}

\item{abundance_weighted}{Boolean, indicates, if abundances should be weighted.}

\item{scaling_iter}{A number of iterations.}

\item{be_verbose}{Boolean, returns messages during the execution.}
}
\value{
A list, result of scaling.
}
\description{
Performs scaling of functional or phylogenetic structure of simulated community.
}
\examples{
com_scaling(com, grain_size = 8, use_tr = T, abundance_weighted = T, scaling_iter = 100, be_verbose = T)
}
