% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_com.r
\name{create_com}
\alias{create_com}
\title{Generates and evolve community under given parameters.}
\usage{
create_com(
  arena_size = 258,
  arena_roughness = 0.75,
  phylo_b = 0.1,
  phylo_d = 0,
  phylo_n_taxa = 500,
  trait_model = "BM",
  trait_sigma = 0.1,
  trait_alpha = 0.1,
  com_range = 3,
  betaEnv = 0,
  betaComp = 0,
  com_nu = 0.001,
  com_sigma = 10,
  sim_n_iter = 2000,
  be_verbose = T,
  sim_verb_iter = 500
)
}
\arguments{
\item{arena_size}{Size of square landscape.}

\item{arena_roughness}{Level of spatial autocorrelation of landscape.}

\item{phylo_b}{Birth rate of phylogenetic tree.}

\item{phylo_d}{Death rate of phylogenetic tree.}

\item{phylo_n_taxa}{A number of species in the tree.}

\item{trait_model}{A model, under which trait should be generated ("BM" or "OU").}

\item{trait_sigma}{A standard-deviation of the random component for each branch in trait simulation.}

\item{trait_alpha}{A strength of the selective constraint for each branch in OU model.}

\item{com_range}{A number, range in which modeling species interact.}

\item{betaEnv}{A number, coefficient for environmental filtering.}

\item{betaComp}{A number, coefficient for limiting similarity.}

\item{com_nu}{A number, rate of speciation.}

\item{com_sigma}{A number, niche width.}

\item{sim_n_iter}{A number of iterations of community simulation.}

\item{be_verbose}{Boolean, returns messages during the execution.}

\item{sim_verb_iter}{A number of iterations in community simulation after which the message is returned.}
}
\value{
A list, community data.
}
\description{
Generates and evolve community under given parameters.
}
\examples{
create_com(arena_size = 258, arena_roughness = 0.75,
phylo_b = 0.1, phylo_d = 0, phylo_n_taxa = 500,
trait_model = "BM", trait_sigma = 0.1, trait_alpha = 0.1,
com_range = 3, betaEnv = 0, betaComp = 0,
com_nu = 0.001, com_sigma = 10,
sim_n_iter = 2000, be_verbose = T, sim_verb_iter = 500)
}
