% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolve_com.r
\name{evolve_com}
\alias{evolve_com}
\title{Simulated previously generated community.}
\usage{
evolve_com(com, n_iter = 2000, be_verbose = T, verb_iter = 500)
}
\arguments{
\item{com}{A list, returned by create_com.}

\item{n_iter}{A number of iterations of simulation.}

\item{be_verbose}{Boolean, returns messages during the execution.}

\item{verb_iter}{A number of iterations in community simulation after which the message is returned.}
}
\value{
A list, community data.
}
\description{
Simulated previously generated community.
}
\examples{
evolve_com(com, n_iter = 2000, be_verbose = T, verb_iter = 500)
}
