% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.cdm.cda.r
\name{make.cda}
\alias{make.cda}
\title{Makes community data array of given size from simulated community.}
\usage{
make.cda(com, size)
}
\arguments{
\item{com}{A list, returned by evolve_com.}

\item{size}{A number, size of each community data matrix.}
}
\value{
Community data array.
}
\description{
Makes community data array of given size from simulated community.
}
\examples{
make.cda(com, size = 8)
}
