% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.tr.dist.r
\name{make.tr.dist}
\alias{make.tr.dist}
\title{Makes matrix of functional distances.}
\usage{
make.tr.dist(com)
}
\arguments{
\item{com}{A list, returned by evolve_com.}
}
\value{
Matrix of functional distances.
}
\description{
Makes matrix of functional distances.
}
\examples{
make.tr.dist(com)
}
