% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDynamicsCom.r
\name{plotCom}
\alias{plotCom}
\title{Visualizes generated community.}
\usage{
plotCom(com)
}
\arguments{
\item{com}{A list, returned by evolve_com.}
}
\description{
Visualizes generated community.
}
\examples{
plotCom(com)
}
