% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDynamicsCom.r
\name{plotDynamics}
\alias{plotDynamics}
\title{Visualizes generated community.}
\usage{
plotDynamics(com, com2 = NULL)
}
\arguments{
\item{com}{A list, returned by evolve_com.}

\item{com2}{A list, returned by evolve_com.}
}
\description{
Visualizes generated community.
}
\examples{
plotDynamics(com, com2 = NULL)
}
