% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_pipeline.r
\name{scaling_pipeline}
\alias{scaling_pipeline}
\title{Performes community modeling and scaling of phylogenetic and functional structure}
\usage{
scaling_pipeline(
  script,
  n_iter,
  sim_verb_iter = 100,
  be_verbose = T,
  make_backup = T,
  use_backup = F,
  backup_path = "backup.rda"
)
}
\arguments{
\item{script}{A Dataframe containing parameters of model and scaling for each script.}

\item{n_iter}{A number of iterations for each script.}

\item{sim_verb_iter}{A number of iterations in community simulation after which the message is returned.}

\item{be_verbose}{Boolean, returns messages during the execution.}

\item{make_backup}{Boolean, indicates if results should be restored in backup file.}

\item{use_backup}{Boolean, use a previous backup after interuption.}

\item{backup_path}{A character, path of backup file.}
}
\value{
List of generated communities and scaling results.
}
\description{
Performes community modeling and scaling of phylogenetic and functional structure
}
\examples{
scaling_pipeline(script, n_iter, sim_verb_iter = 100, be_verbose = T,
make_backup = T, use_backup = F, backup_path = "backup.rda")
}
