% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_draft.r
\name{script_draft}
\alias{script_draft}
\title{Supporting function for scaling_pipeline. Creates a Dataframe containing all parameters in model.}
\usage{
script_draft(n)
}
\arguments{
\item{n}{A number of rows.}
}
\value{
Dataframe.
}
\description{
Supporting function for scaling_pipeline. Creates a Dataframe containing all parameters in model.
}
\examples{
script_draft(8)
}
