% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_arena.r
\name{sim_arena}
\alias{sim_arena}
\title{Creates square landscape on which community will be simulated.}
\usage{
sim_arena(size = 256, roughness = 0.75)
}
\arguments{
\item{size}{Size of square landscape.}

\item{roughness}{Level of spatial autocorrelation.}
}
\value{
Landscape
}
\description{
Creates square landscape on which community will be simulated.
}
\examples{
sim_arena(size = 256, roughness = 0.75)
}
