% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_tree_n_traits.r
\name{sim_tree_n_traits}
\alias{sim_tree_n_traits}
\title{Creates phylogenetic tree and generates a single trait for each species.}
\usage{
sim_tree_n_traits(
  b = 0.1,
  d = 0,
  n_taxa = 500,
  model = "BM",
  sigma = 0.1,
  alpha = 0.1
)
}
\arguments{
\item{b}{Birth rate.}

\item{d}{Death rate.}

\item{n_taxa}{A number of species in the tree.}

\item{model}{A model, under which trait should be generated ("BM" or "OU").}

\item{sigma}{A standard-deviation of the random component for each branch.}

\item{alpha}{A strength of the selective constraint for each branch in OU model.}
}
\value{
List consisting of phylogenetic tree, matrix of phylogenetic distances, trait and matrix of trait distances.
}
\description{
Creates phylogenetic tree and generates a single trait for each species.
}
\examples{
sim_tree_n_traits(b = 0.1, d = 0, n_taxa = 500, model = "BM", sigma = 0.1, alpha = 0.1)
}
